+++ resources/lang/ca/students.php
@@ -53,5 +53,5 @@ return [
     'highlights' => 'Destaca en',
     'change_class' => 'Canvia subclasse',
     'basic_equipment' => 'L\'equipament bàsic serà assignat, aquesta acció no es pot desfer.',
-    '' => '',
+    'online' => 'L\'estudiant està actiu (online)',
 ];